RunSwig.batをいじったら、リンクエラーが出るようになってしまったので戻しました。
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@4612 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Graphics/GRFrame.cpp
@@ -371,7 +371,7 @@ bool GRAnimation::AddChildObject(ObjectIf* o){
 }
 
 GRAnimationKey GRAnimation::GetAnimationKey(int n){
-	if(n < keys.size()){
+	if(n < (int)keys.size()){
 		return keys[n];
 	}else{
 		DSTR<<"GetAnimationKey is NO"<<std::endl;
@@ -515,13 +515,13 @@ void GRAnimationSet::LoadInitialPose(){
 }
 
 void GRAnimationSet::SetCurrentAnimationPose(float t){
-	for(int i = 0; i<animations.size(); i++){
+	for(size_t i = 0; i<animations.size(); i++){
 		animations[i]->SetCurrentPose(t);
 	}
 	UpdateLastKeyTime();
 }
 void GRAnimationSet::DeleteAnimationPose(float t){
-	for(int i = 0; i<animations.size(); i++){
+	for(size_t i = 0; i<animations.size(); i++){
 		animations[i]->DeletePose(t);
 	}
 	UpdateLastKeyTime();
